+++ demos/python/sdk_wireless_camera_control/docs/changelog.rst
@@ -9,6 +9,11 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+Unreleased
+----------
+
+* Add Setting 125
+
 0.16.1 (April-23-2024)
 ----------------------