+++ demos/python/sdk_wireless_camera_control/docs/changelog.rst
@@ -9,8 +9,8 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-Unreleased
-----------
+0.16.2 (July-18-2024)
+---------------------
 
 * Add Setting 125
 * Don't default to hardcoded parameters for set livestream mode